refactor(portal): move shared notification strings to notification service (bgu-nasa/main#142)
