+++ src/modules/resources/src/pages/activities-page/activities-page.tsx
@@ -18,7 +18,7 @@ import resourcesJson from "./activities-page.resources.json";
 import styles from "./activities-page.module.css";
 import { userRepository } from "@/modules/resources/src/data";
 import { translatedResources } from "@/infra/i18n";
-import { sharedNotifications } from "@/infra/i18n/shared-notifications";
+import { sharedNotifications } from "@/infra/service/notification";
 
 const resources = translatedResources("src/modules/resources/src/pages/activities-page/activities-page.resources.json", resourcesJson);
