+++ src/modules/resources/src/pages/subjects-page/components/subject-creator.tsx
@@ -3,7 +3,7 @@ import { useState, useEffect } from "react";
 import { schedulingPeriodRepository } from "@/modules/resources/src/data";
 import resourcesJson from "./subject-creator.resources.json";
 import { translatedResources } from "@/infra/i18n";
-import { sharedNotifications } from "@/infra/i18n/shared-notifications";
+import { sharedNotifications } from "@/infra/service/notification";
 
 const resources = translatedResources(
     "src/modules/resources/src/pages/subjects-page/components/subject-creator.resources.json",
