+++ src/modules/resources/src/pages/subjects-page/subjects-page.tsx
@@ -19,7 +19,7 @@ import resourcesJson from "./subjects-page.resources.json";
 import styles from "./subjects-page.module.css";
 import { schedulingPeriodRepository, departmentRepository } from "@/modules/resources/src/data";
 import { translatedResources } from "@/infra/i18n";
-import { sharedNotifications } from "@/infra/i18n/shared-notifications";
+import { sharedNotifications } from "@/infra/service/notification";
 
 const resources = translatedResources("src/modules/resources/src/pages/subjects-page/subjects-page.resources.json", resourcesJson);
